requirements.txt
Django==3.2.23
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
esdk-obs-python==3.22.2
icalendar==4.1.0
mysqlclient==2.0.3
PyMySQL==1.1.0
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0
uwsgi==2.0.22